refactor(kimi-k2): extract KimiK2DirectRuntimeConfig and drop dead fields
scheduler.rs was carrying the runtime config alongside the scheduler /
runtime impls, plus two genuinely dead config fields. Move the struct
into a focused `direct/config.rs`, delete what no one reads.

- `use_nccl_ep_bridge`: only ever constructed as `true`, only read once
  in `KimiK2DirectRuntime::spawn` to bail with "PPLX EP not wired yet"
  if false. Nothing can set it false, so the field, the assignment, and
  the guard are all unreachable. Delete.

- `text_config`: loaded from config.json by `load_text_config`, stored
  in the runtime config, then never read by anyone. The loader function
  itself is dead too. Delete both, and stop importing `KimiK2TextConfig`
  / `probe_config_json` from this file. `lib.rs` still re-exports those
  for callers that actually need them.

Also: collectives.rs's module doc pointed at the deleted
`use_nccl_ep_bridge` field; rewrite it to describe the bridge by its
real call sites instead of by a flag that no longer exists.

scheduler.rs: 757 → 719 lines. Public surface is unchanged —
`KimiK2DirectRuntimeConfig` is still re-exported from `direct` and
`lib`, only its definition site moved.
