+++ pegainfer-kimi-k2/src/direct.rs
@@ -1,8 +1,9 @@
 mod affinity;
+mod config;
 mod scheduler;
 mod worker;
 
-pub use scheduler::KimiK2DirectRuntimeConfig;
+pub use config::KimiK2DirectRuntimeConfig;
 pub use worker::KimiK2RankPlacement;
 
 pub(crate) use scheduler::start_engine;
